requirements.txt
python-dotenv
pymongo
ambient_api